Removed Regions.
